+++ IOFireWireFamily/IOFireWireLib.CFPlugInProj/IOFireWireLibAsyncStreamListener.cpp
@@ -17,93 +17,93 @@
 #endif
 
 namespace IOFireWireLib {
-
+	
 #pragma mark AsyncStreamListener -
-
+	
 	// ============================================================
 	// AsyncStreamListener
 	// ============================================================
-	AsyncStreamListener::AsyncStreamListener( const IUnknownVTbl&		interface, 
-													Device&				userclient,
-													UserObjectHandle	inKernAddrSpaceRef,
-													void*				inBuffer,
-													UInt32				inBufferSize,
-													void*				inCallBack,
-													void*				inRefCon	)
-	: IOFireWireIUnknown( interface ), 
-		mUserClient(userclient), 
-		mKernAsyncStreamListenerRef(inKernAddrSpaceRef),
-		mBuffer((char*)inBuffer),
-		mNotifyIsOn(false),
-		mUserRefCon(inRefCon),
-		mBufferSize(inBufferSize),
-		mListener( (AsyncStreamListenerHandler) inCallBack ),
-		mSkippedPacketHandler( nil ),
-		mRefInterface( reinterpret_cast<AsyncStreamListenerRef>( & GetInterface() ) )
+	AsyncStreamListener::AsyncStreamListener( const IUnknownVTbl&		interface,
+											 Device&				userclient,
+											 UserObjectHandle	inKernAddrSpaceRef,
+											 void*				inBuffer,
+											 UInt32				inBufferSize,
+											 void*				inCallBack,
+											 void*				inRefCon	)
+	: IOFireWireIUnknown( interface ),
+	mUserClient(userclient),
+	mKernAsyncStreamListenerRef(inKernAddrSpaceRef),
+	mBuffer((char*)inBuffer),
+	mNotifyIsOn(false),
+	mUserRefCon(inRefCon),
+	mBufferSize(inBufferSize),
+	mListener( (AsyncStreamListenerHandler) inCallBack ),
+	mSkippedPacketHandler( nil ),
+	mRefInterface( reinterpret_cast<AsyncStreamListenerRef>( & GetInterface() ) )
 	{
 		userclient.AddRef() ;
 	}
 	
-
+	
 	AsyncStreamListener::~AsyncStreamListener()
 	{
 		uint32_t outputCnt = 0;
 		const uint64_t inputs[1]={(const uint64_t)mKernAsyncStreamListenerRef};
-
-		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(), 
+		
+		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(),
 								  kReleaseUserObject,
 								  inputs,1,
 								  NULL,&outputCnt);
 		
-		if( mBuffer and mBufferSize > 0 )	
+		if( mBuffer and mBufferSize > 0 )
 		{
 			delete[] mBuffer;
 			mBuffer		= 0;
 			mBufferSize = 0;
 		}
-			
+		
 		mUserClient.Release() ;
 	}
-
-	const IOFWAsyncStreamListenerHandler 
-	AsyncStreamListener::SetListenerHandler ( AsyncStreamListenerRef		self, 
-											  AsyncStreamListenerHandler	inReceiver )
+	
+	const IOFWAsyncStreamListenerHandler
+	AsyncStreamListener::SetListenerHandler ( AsyncStreamListenerRef		self,
+											 AsyncStreamListenerHandler	inReceiver )
 	{
 		AsyncStreamListenerHandler oldListener = mListener ;
 		mListener = inReceiver ;
 		
 		return oldListener ;
 	}
-
-	const IOFWAsyncStreamListenerSkippedPacketHandler	
-	AsyncStreamListener::SetSkippedPacketHandler( AsyncStreamListenerRef	self, 
-												  AsyncStreamSkippedPacketHandler		inHandler )
+	
+	const IOFWAsyncStreamListenerSkippedPacketHandler
+	AsyncStreamListener::SetSkippedPacketHandler( AsyncStreamListenerRef	self,
+												 AsyncStreamSkippedPacketHandler		inHandler )
 	{
 		AsyncStreamSkippedPacketHandler oldHandler = mSkippedPacketHandler;
 		mSkippedPacketHandler = inHandler;
 		
 		return oldHandler;
 	}
-
-	Boolean 
-	AsyncStreamListener::NotificationIsOn ( AsyncStreamListenerRef self ) 
+	
+	Boolean
+	AsyncStreamListener::NotificationIsOn ( AsyncStreamListenerRef self )
 	{
 		return mNotifyIsOn ;
 	}
-
-	Boolean 
+	
+	Boolean
 	AsyncStreamListener::TurnOnNotification ( AsyncStreamListenerRef self )
 	{
 		IOReturn				err					= kIOReturnSuccess ;
 		io_connect_t			connection			= mUserClient.GetUserClientConnection() ;
-	
+		
 		// if notification is already on, skip out.
 		if (mNotifyIsOn)
 			return true ;
 		
 		if (!connection)
 			err = kIOReturnNoDevice ;
-
+		
 		if ( kIOReturnSuccess == err )
 		{
 			uint64_t refrncData[kOSAsyncRef64Count];
@@ -121,7 +121,7 @@ namespace IOFireWireLib {
 		
 		if ( kIOReturnSuccess == err)
 		{
-	
+			
 			uint64_t refrncData[kOSAsyncRef64Count];
 			refrncData[kIOAsyncCalloutFuncIndex] = (uint64_t) 0;
 			refrncData[kIOAsyncCalloutRefconIndex] = (unsigned long) 0;
@@ -134,38 +134,38 @@ namespace IOFireWireLib {
 												 inputs,3,
 												 NULL,&outputCnt);
 		}
-
+		
 		if ( kIOReturnSuccess == err )
 		{
 			uint32_t outputCnt = 0;
 			const uint64_t inputs[1]={(const uint64_t)mKernAsyncStreamListenerRef};
-
-			err = IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(), 
+			
+			err = IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(),
 											kAsyncStreamListener_TurnOnNotification,
 											inputs,1,
 											NULL,&outputCnt);
 		}
-
+		
 		
 		if ( kIOReturnSuccess == err )
 			mNotifyIsOn = true ;
-			
+		
 		return ( kIOReturnSuccess == err ) ;
 	}
-
-	void 
+	
+	void
 	AsyncStreamListener::TurnOffNotification ( AsyncStreamListenerRef self )
 	{
 		IOReturn				err			= kIOReturnSuccess ;
 		io_connect_t			connection	= mUserClient.GetUserClientConnection() ;
-
+		
 		// if notification isn't on, skip out.
 		if (!mNotifyIsOn)
 			return ;
-	
+		
 		if (!connection)
 			err = kIOReturnNoDevice ;
-	
+		
 		if ( kIOReturnSuccess == err )
 		{
 			uint64_t refrncData[kOSAsyncRef64Count];
@@ -173,7 +173,7 @@ namespace IOFireWireLib {
 			refrncData[kIOAsyncCalloutRefconIndex] = (unsigned long) 0;
 			const uint64_t inputs[3] = {(const uint64_t)mKernAsyncStreamListenerRef,0,(const uint64_t)self};
 			uint32_t outputCnt = 0;
-
+			
 			// set callback for writes to 0
 			err = IOConnectCallAsyncScalarMethod(connection,
 												 kSetAsyncStreamRef_Packet,
@@ -181,7 +181,7 @@ namespace IOFireWireLib {
 												 refrncData,kOSAsyncRef64Count,
 												 inputs,3,
 												 NULL,&outputCnt);
-
+			
 			outputCnt = 0;
 			// set callback for skipped packets to 0
 			err = IOConnectCallAsyncScalarMethod(connection,
@@ -191,13 +191,13 @@ namespace IOFireWireLib {
 												 inputs,3,
 												 NULL,&outputCnt);
 		}
-
+		
 		if ( kIOReturnSuccess == err )
 		{
 			uint32_t outputCnt = 0;
 			const uint64_t inputs[1]={(const uint64_t)mKernAsyncStreamListenerRef};
-
-			err = IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(), 
+			
+			err = IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(),
 											kAsyncStreamListener_TurnOffNotification,
 											inputs,1,
 											NULL,&outputCnt);
@@ -205,22 +205,22 @@ namespace IOFireWireLib {
 		
 		mNotifyIsOn = false ;
 	}
-
-	void 
+	
+	void
 	AsyncStreamListener::ClientCommandIsComplete ( AsyncStreamListenerRef	self,
-												   FWClientCommandID		commandID )
+												  FWClientCommandID		commandID )
 	{
 		uint32_t		outputCnt = 0;
 		const uint64_t	inputs[2] = {(const uint64_t)mKernAsyncStreamListenerRef, (const uint64_t)commandID};
-
-		#if IOFIREWIREUSERCLIENTDEBUG > 0
-		OSStatus err = 
-		#endif
-
-		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(), 
-													kAsyncStreamListener_ClientCommandIsComplete,
-													inputs,2,
-													NULL,&outputCnt);
+		
+#if IOFIREWIREUSERCLIENTDEBUG > 0
+		OSStatus err =
+#endif
+		
+		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(),
+								  kAsyncStreamListener_ClientCommandIsComplete,
+								  inputs,2,
+								  NULL,&outputCnt);
 		
 #ifdef __LP64__
 		DebugLogCond( err, "AsyncStreamListener::ClientCommandIsComplete: err=0x%08X\n", err ) ;
@@ -228,35 +228,35 @@ namespace IOFireWireLib {
 		DebugLogCond( err, "AsyncStreamListener::ClientCommandIsComplete: err=0x%08lX\n", err ) ;
 #endif
 	}
-
-	void* 
-	AsyncStreamListener::GetRefCon	( AsyncStreamListenerRef self )	
+	
+	void*
+	AsyncStreamListener::GetRefCon	( AsyncStreamListenerRef self )
 	{
 		return mUserRefCon;
 	}
-
-	void 
+	
+	void
 	AsyncStreamListener::SetFlags ( AsyncStreamListenerRef		self,
-									UInt32						flags )
+								   UInt32						flags )
 	{
 		uint32_t outputCnt = 0;
 		const uint64_t inputs[2] = {(const uint64_t)mKernAsyncStreamListenerRef, flags};
-		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(), 
+		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(),
 								  kAsyncStreamListener_SetFlags,
 								  inputs,2,
 								  NULL,&outputCnt);
 		
 		mFlags = flags;
 	}
-		
-	UInt32 
-	AsyncStreamListener::GetFlags ( AsyncStreamListenerRef	self )	
+	
+	UInt32
+	AsyncStreamListener::GetFlags ( AsyncStreamListenerRef	self )
 	{
 		uint32_t outputCnt = 1;
 		uint64_t outputVal = 0;
 		const uint64_t inputs[1]={(const uint64_t)mKernAsyncStreamListenerRef};
-
-		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(), 
+		
+		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(),
 								  kAsyncStreamListener_GetFlags,
 								  inputs,1,
 								  &outputVal,&outputCnt);
@@ -264,7 +264,7 @@ namespace IOFireWireLib {
 		
 		return mFlags;
 	}
-
+	
 	UInt32
 	AsyncStreamListener::GetOverrunCounter ( AsyncStreamListenerRef	self )
 	{
@@ -272,15 +272,15 @@ namespace IOFireWireLib {
 		uint32_t outputCnt = 1;
 		uint64_t outputVal = 0;
 		const uint64_t inputs[1]={(const uint64_t)mKernAsyncStreamListenerRef};
-
-		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(), 
+		
+		IOConnectCallScalarMethod(mUserClient.GetUserClientConnection(),
 								  kAsyncStreamListener_GetOverrunCounter,
 								  inputs,1,
 								  &outputVal,&outputCnt);
 		counter = outputVal & 0xFFFFFFFF;
 		return counter;
 	}
-
+	
 	void
 	AsyncStreamListener::Listener( AsyncStreamListenerRef refcon, IOReturn result, void** args, int numArgs)
 	{
@@ -294,11 +294,11 @@ namespace IOFireWireLib {
 		else
 		{
 			(me->mListener)(
-				(AsyncStreamListenerRef) refcon,
-				(FWClientCommandID) args[0],						// commandID,
-				(unsigned long)(args[1]),									// size
-				me->mBuffer + (unsigned long)(args[2]),					// packet
-				(void*) me->mUserRefCon) ;								// refcon
+							(AsyncStreamListenerRef) refcon,
+							(FWClientCommandID)args[0],						// commandID,
+							(UInt32)((unsigned long)args[1]),									// size
+							me->mBuffer + (unsigned long)args[2],					// packet
+							(void*) me->mUserRefCon) ;								// refcon
 		}
 	}
 	
@@ -306,17 +306,17 @@ namespace IOFireWireLib {
 	AsyncStreamListener::SkippedPacket( AsyncStreamListenerRef refcon, IOReturn result, FWClientCommandID commandID, UInt32 packetCount)
 	{
 		AsyncStreamListener* me = IOFireWireIUnknown::InterfaceMap<AsyncStreamListener>::GetThis(refcon) ;
-	
+		
 		if (me->mSkippedPacketHandler)
 			(me->mSkippedPacketHandler)( refcon, commandID, packetCount) ;
 		else
 			me->ClientCommandIsComplete( refcon, commandID ) ;
 	}
-
-
+	
+	
 #pragma mark AsyncStreamListenerCOM -
-
-	AsyncStreamListenerCOM::Interface AsyncStreamListenerCOM::sInterface = 
+	
+	AsyncStreamListenerCOM::Interface AsyncStreamListenerCOM::sInterface =
 	{
 		INTERFACEIMP_INTERFACE,
 		1, 0,
@@ -325,7 +325,7 @@ namespace IOFireWireLib {
 		&AsyncStreamListenerCOM::SSetSkippedPacketHandler,
 		&AsyncStreamListenerCOM::SNotificationIsOn,
 		&AsyncStreamListenerCOM::STurnOnNotification,
-		&AsyncStreamListenerCOM::STurnOffNotification, 
+		&AsyncStreamListenerCOM::STurnOffNotification,
 		&AsyncStreamListenerCOM::SClientCommandIsComplete,
 		&AsyncStreamListenerCOM::SGetRefCon,
 		&AsyncStreamListenerCOM::SSetFlags,
@@ -338,13 +338,13 @@ namespace IOFireWireLib {
 	//
 	
 	AsyncStreamListenerCOM::AsyncStreamListenerCOM( Device&				userclient,
-													UserObjectHandle	inKernAddrSpaceRef,
-													void*				inBuffer,
-													UInt32				inBufferSize,
-													void*				inCallBack,
-													void*				inRefCon )
-	: AsyncStreamListener( reinterpret_cast<const IUnknownVTbl &>( sInterface ), userclient, inKernAddrSpaceRef, inBuffer, 
-										inBufferSize, inCallBack, inRefCon )
+												   UserObjectHandle	inKernAddrSpaceRef,
+												   void*				inBuffer,
+												   UInt32				inBufferSize,
+												   void*				inCallBack,
+												   void*				inRefCon )
+	: AsyncStreamListener( reinterpret_cast<const IUnknownVTbl &>( sInterface ), userclient, inKernAddrSpaceRef, inBuffer,
+	inBufferSize, inCallBack, inRefCon )
 	{
 	}
 	
@@ -355,14 +355,14 @@ namespace IOFireWireLib {
 	//
 	// --- IUNKNOWN support ----------------
 	//
-		
+	
 	IUnknownVTbl**
 	AsyncStreamListenerCOM::Alloc(	Device&				userclient,
-									UserObjectHandle	inKernAddrSpaceRef,
-									void*				inBuffer,
-									UInt32				inBufferSize,
-									void*				inCallBack,
-									void*				inRefCon )
+								  UserObjectHandle	inKernAddrSpaceRef,
+								  void*				inBuffer,
+								  UInt32				inBufferSize,
+								  void*				inCallBack,
+								  void*				inRefCon )
 	{
 		AsyncStreamListenerCOM*	me = nil ;
 		
@@ -370,7 +370,7 @@ namespace IOFireWireLib {
 			me = new AsyncStreamListenerCOM( userclient, inKernAddrSpaceRef, inBuffer, inBufferSize, inCallBack, inRefCon ) ;
 		} catch(...) {
 		}
-
+		
 		return ( nil == me ) ? nil : reinterpret_cast<IUnknownVTbl**>(& me->GetInterface()) ;
 	}
 	
@@ -379,9 +379,9 @@ namespace IOFireWireLib {
 	{
 		HRESULT		result = S_OK ;
 		*ppv = nil ;
-	
+		
 		CFUUIDRef	interfaceID	= CFUUIDCreateFromUUIDBytes(kCFAllocatorDefault, iid) ;
-	
+		
 		if ( CFEqual(interfaceID, IUnknownUUID) ||  CFEqual(interfaceID, kIOFireWireAsyncStreamListenerInterfaceID) )
 		{
 			*ppv = & GetInterface() ;
@@ -391,7 +391,7 @@ namespace IOFireWireLib {
 		{
 			*ppv = nil ;
 			result = E_NOINTERFACE ;
-		}	
+		}
 		
 		CFRelease(interfaceID) ;
 		return result ;
@@ -400,68 +400,69 @@ namespace IOFireWireLib {
 	//
 	// --- static methods ------------------
 	//
-	const IOFWAsyncStreamListenerHandler 
-	AsyncStreamListenerCOM::SSetListenerHandler ( AsyncStreamListenerRef		self, 
-												  AsyncStreamListenerHandler	inReceiver )
+	const IOFWAsyncStreamListenerHandler
+	AsyncStreamListenerCOM::SSetListenerHandler ( AsyncStreamListenerRef		self,
+												 AsyncStreamListenerHandler	inReceiver )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->SetListenerHandler( self, inReceiver ) ;
 	}
-
-	const IOFWAsyncStreamListenerSkippedPacketHandler	
-	AsyncStreamListenerCOM::SSetSkippedPacketHandler( AsyncStreamListenerRef	self, 
-													  AsyncStreamSkippedPacketHandler		inHandler )
+	
+	const IOFWAsyncStreamListenerSkippedPacketHandler
+	AsyncStreamListenerCOM::SSetSkippedPacketHandler( AsyncStreamListenerRef	self,
+													 AsyncStreamSkippedPacketHandler		inHandler )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->SetSkippedPacketHandler( self, inHandler ) ;
 	}
-
-	Boolean 
-	AsyncStreamListenerCOM::SNotificationIsOn ( AsyncStreamListenerRef self ) 
+	
+	Boolean
+	AsyncStreamListenerCOM::SNotificationIsOn ( AsyncStreamListenerRef self )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->NotificationIsOn( self ) ;
 	}
-
-	Boolean 
+	
+	Boolean
 	AsyncStreamListenerCOM::STurnOnNotification ( AsyncStreamListenerRef self )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->TurnOnNotification( self ) ;
 	}
-
-	void 
+	
+	void
 	AsyncStreamListenerCOM::STurnOffNotification ( AsyncStreamListenerRef self )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->TurnOffNotification( self ) ;
 	}
-
-	void 
+	
+	void
 	AsyncStreamListenerCOM::SClientCommandIsComplete ( AsyncStreamListenerRef	self,
-													   FWClientCommandID		commandID,
-													   IOReturn					status )
+													  FWClientCommandID		commandID,
+													  IOReturn					status )
 	{
 		IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->ClientCommandIsComplete( self, commandID ) ;
 	}
-
-	void* 
-	AsyncStreamListenerCOM::SGetRefCon	( AsyncStreamListenerRef self )	
+	
+	void*
+	AsyncStreamListenerCOM::SGetRefCon	( AsyncStreamListenerRef self )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->GetRefCon( self );
 	}
-
+	
 	void
 	AsyncStreamListenerCOM::SSetFlags ( AsyncStreamListenerRef		self,
-										 UInt32						flags )
+									   UInt32						flags )
 	{
 		IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->SetFlags( self, flags );
 	}
-		
+	
 	UInt32
-	AsyncStreamListenerCOM::SGetFlags ( AsyncStreamListenerRef	self )	
+	AsyncStreamListenerCOM::SGetFlags ( AsyncStreamListenerRef	self )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->GetFlags( self );
 	}
-
+	
 	UInt32
 	AsyncStreamListenerCOM::SGetOverrunCounter ( AsyncStreamListenerRef		self )
 	{
 		return IOFireWireIUnknown::InterfaceMap<AsyncStreamListenerCOM>::GetThis(self)->GetOverrunCounter( self );
 	}
 }
+
